do git reset in order to delete unnecessary changes in this PR
