+++ backend/src/helpers/mailtrap/templates/reset-password-request.ts
@@ -1,0 +1,56 @@
+import { ENV } from '@common/enums/app/env.enum';
+
+import { NodemailerPayload } from '../interfaces/nodemailer-payload';
+
+export const getMessage = (payload: NodemailerPayload): string => {
+  return `
+    <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+    <html xmlns="http://www.w3.org/1999/xhtml">
+      <head>
+        <title>HTML email template</title>
+        <meta name="viewport" content="width = 375, initial-scale = -1">
+      </head>
+
+      <style>
+        html {
+          font-family: Lato, sans-serif;
+        }
+      </style>
+
+      <body style="background-color: #ffffff; font-size: 16px;">
+        <center>
+          <table align="center" border="0" cellpadding="0" cellspacing="0" style="height:100%; width:600px;">
+              <tr>
+                <td align="center" bgcolor="#ffffff" style="padding:30px">
+                  <p>
+                    <svg width="101" height="30" viewBox="0 0 101 30" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <path d="M98.1114 6.36532C102.461 14.9209 100.047 21.4029 98.2966 23.5744C101.456 14.8781 98.2966 10.3666 97.296 8.31553C87.5615 11.3272 81.2325 9.62749 74.8271 8.31553C69.1162 6.84922 64.2542 3.91661 51.3662 2.68182C38.4781 1.44704 19.1111 6.77776 17.0651 6.81788C20.8762 5.49401 27.3715 3.3281 37.8888 0.956536C48.4062 -1.41503 61.2701 0.843903 73.5503 5.3039C85.8306 9.76389 95.7513 6.96281 98.1114 6.36532Z" fill="#3765FF"/>
+                    <path d="M9.20316 23.1367L7.98756 19.544C7.89751 19.3159 7.80147 19.0487 7.69942 18.7426C7.59737 18.4304 7.49532 18.0943 7.39327 17.7341C7.30322 18.1003 7.20717 18.4394 7.10512 18.7516C7.00307 19.0637 6.90702 19.3339 6.81698 19.562L5.61038 23.1367H9.20316ZM14.0385 28.1522H11.6794C11.4153 28.1522 11.2021 28.0922 11.0401 27.9721C10.878 27.8461 10.7549 27.687 10.6709 27.4949L9.8965 25.2078H4.90804L4.13365 27.4949C4.06762 27.663 3.94756 27.8161 3.77348 27.9541C3.60539 28.0862 3.39529 28.1522 3.14317 28.1522H0.765991L5.8445 15.0327H8.96004L14.0385 28.1522Z" fill="#070A0C"/>
+                    <path d="M20.2135 25.748C20.6217 25.748 20.9849 25.682 21.303 25.5499C21.6272 25.4119 21.9003 25.2168 22.1224 24.9646C22.3445 24.7125 22.5126 24.4064 22.6267 24.0462C22.7467 23.68 22.8068 23.2658 22.8068 22.8036V15.0327H25.8503V22.8036C25.8503 23.608 25.7182 24.3463 25.4541 25.0187C25.1959 25.691 24.8238 26.2703 24.3375 26.7565C23.8513 27.2428 23.26 27.621 22.5636 27.8911C21.8673 28.1612 21.0839 28.2963 20.2135 28.2963C19.3371 28.2963 18.5507 28.1612 17.8543 27.8911C17.158 27.621 16.5667 27.2428 16.0804 26.7565C15.5942 26.2703 15.222 25.691 14.9639 25.0187C14.7058 24.3463 14.5767 23.608 14.5767 22.8036V15.0327H17.6202V22.7946C17.6202 23.2568 17.6772 23.671 17.7913 24.0372C17.9114 24.3974 18.0824 24.7065 18.3045 24.9646C18.5267 25.2168 18.7968 25.4119 19.1149 25.5499C19.4391 25.682 19.8053 25.748 20.2135 25.748Z" fill="#070A0C"/>
+                    <path d="M37.6555 17.4459H33.9637V28.1522H30.9202V17.4459H27.2284V15.0327H37.6555V17.4459Z" fill="#070A0C"/>
+                    <path d="M51.7794 21.588C51.7794 22.5364 51.6143 23.4189 51.2842 24.2353C50.96 25.0517 50.5008 25.763 49.9065 26.3693C49.3122 26.9696 48.5949 27.4409 47.7544 27.783C46.92 28.1252 45.9926 28.2963 44.9721 28.2963C43.9516 28.2963 43.0211 28.1252 42.1807 27.783C41.3403 27.4409 40.6199 26.9696 40.0196 26.3693C39.4253 25.763 38.9631 25.0517 38.6329 24.2353C38.3088 23.4189 38.1467 22.5364 38.1467 21.588C38.1467 20.6395 38.3088 19.7571 38.6329 18.9407C38.9631 18.1243 39.4253 17.4159 40.0196 16.8156C40.6199 16.2153 41.3403 15.7441 42.1807 15.4019C43.0211 15.0598 43.9516 14.8887 44.9721 14.8887C45.9926 14.8887 46.92 15.0628 47.7544 15.4109C48.5949 15.7531 49.3122 16.2243 49.9065 16.8246C50.5008 17.4249 50.96 18.1333 51.2842 18.9497C51.6143 19.7661 51.7794 20.6455 51.7794 21.588ZM48.6639 21.588C48.6639 20.9397 48.5798 20.3574 48.4118 19.8411C48.2437 19.3189 48.0006 18.8776 47.6824 18.5175C47.3702 18.1573 46.9861 17.8812 46.5298 17.6891C46.0736 17.497 45.5544 17.4009 44.9721 17.4009C44.3838 17.4009 43.8585 17.497 43.3963 17.6891C42.9401 17.8812 42.5529 18.1573 42.2347 18.5175C41.9226 18.8776 41.6824 19.3189 41.5144 19.8411C41.3463 20.3574 41.2622 20.9397 41.2622 21.588C41.2622 22.2423 41.3463 22.8306 41.5144 23.3529C41.6824 23.8691 41.9226 24.3073 42.2347 24.6675C42.5529 25.0277 42.9401 25.3038 43.3963 25.4959C43.8585 25.682 44.3838 25.775 44.9721 25.775C45.5544 25.775 46.0736 25.682 46.5298 25.4959C46.9861 25.3038 47.3702 25.0277 47.6824 24.6675C48.0006 24.3073 48.2437 23.8691 48.4118 23.3529C48.5798 22.8306 48.6639 22.2423 48.6639 21.588Z" fill="#070A0C"/>
+                    <path d="M61.484 25.721V28.1522H53.5781V15.0327H56.6216V25.721H61.484Z" fill="#3765FF"/>
+                    <path d="M66.1013 28.1522H63.0398V15.0327H66.1013V28.1522Z" fill="#3765FF"/>
+                    <path d="M80.0861 15.0327V28.1522H78.5013C78.2672 28.1522 78.0691 28.1162 77.907 28.0441C77.7509 27.9661 77.5949 27.834 77.4388 27.6479L71.2527 19.8231C71.2767 20.0572 71.2918 20.2853 71.2978 20.5074C71.3098 20.7235 71.3158 20.9276 71.3158 21.1197V28.1522H68.6324V15.0327H70.2352C70.3673 15.0327 70.4783 15.0387 70.5684 15.0507C70.6584 15.0627 70.7395 15.0867 70.8115 15.1228C70.8835 15.1528 70.9526 15.1978 71.0186 15.2578C71.0847 15.3179 71.1597 15.3989 71.2437 15.5009L77.4838 23.3798C77.4538 23.1277 77.4328 22.8846 77.4208 22.6505C77.4088 22.4104 77.4028 22.1852 77.4028 21.9751V15.0327H80.0861Z" fill="#3765FF"/>
+                    <path d="M85.6931 17.3739V20.4354H89.8171V22.6865H85.6931V25.811H91.0777V28.1522H82.6315V15.0327H91.0777V17.3739H85.6931Z" fill="#3765FF"/>
+                    </svg>
+                  </p>
+                  <p style="text-align:left">Hello, ${payload.name}<br><br> We received a request to reset the password for your account for this email address. To initiate the password reset process for your account, click the link below.
+                  </p>
+                  <p>
+                    <a target="_blank" style="text-decoration:none; background-color: #3765FF; border: #3765FF 1px solid; border-radius: 30px; color: #fff; padding:10px 10px; display:block;" href="http://${payload.link}">
+                      <strong>Reset Password</strong></a>
+                  </p>
+                  <p style="text-align:left">This link can only be used once. If you need to reset your password again, please visit <a href="${ENV.APP.FRONTEND_URL}/forgot-password">autoline.com</a> and request another reset.<br><br>If you did not make this request, you can simply ignore this email.</p>
+                  <p style="text-align:left">
+                  Sincerely,<br>Autoline Team
+                  </p>
+                </td>
+              </tr>
+            </tbody>
+          </table>
+        </center>
+      </body>
+    </html>
+  `;
+};
