+++ frontend/src/assets/css/form.scss
@@ -14,6 +14,10 @@
   color: var(--red);
 }
 
+.MuiAlert-standard {
+  margin-bottom: 10px;
+}
+
 .MuiInputBase-root {
   .MuiSelect-select {
     box-sizing: border-box;
